Adding all files of the last updated version of the modpack.
This was the last updated version of the modpack up until I stopped working on it. These are directly from curseforge so they should be a direct drag and drop into your Instances folder. Simply put the folder titled "Dimensions and Dungeoneering" into your "Instances" folder in curseforge and it should work.